added the custom loss function in create_and_compile_model and also did changes in test module.
