+++ skforecast/deep_learning/utils.py
@@ -257,87 +257,3 @@ def create_and_compile_model(
         model.compile(optimizer=optimizer, loss=loss, **compile_kwargs)
 
     return model
-
-
-# Unit tests
-def test_predefined_loss_string():
-    """Test model creation with predefined loss string"""
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    model = create_and_compile_model(
-        series=data,
-        lags=2,
-        steps=1,
-        loss='mse'
-    )
-    assert model.loss == 'mse'
-
-def test_predefined_loss_object():
-    """Test model creation with predefined loss object"""
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    model = create_and_compile_model(
-        series=data,
-        lags=2,
-        steps=1,
-        loss=MeanSquaredError()
-    )
-    assert isinstance(model.loss, MeanSquaredError)
-
-def test_custom_loss_function():
-    """Test model creation with custom loss function"""
-    def custom_loss(y_true, y_pred):
-        return tf.reduce_mean(tf.square(y_true - y_pred))
-    
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    model = create_and_compile_model(
-        series=data,
-        lags=2,
-        steps=1,
-        loss=custom_loss
-    )
-    assert model.loss == custom_loss
-
-def test_invalid_custom_loss():
-    """Test that invalid custom loss raises error"""
-    def invalid_loss(y_true, y_pred, extra_param):
-        return tf.reduce_mean(tf.square(y_true - y_pred))
-    
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    with pytest.raises(ValueError) as exc_info:
-        create_and_compile_model(
-            series=data,
-            lags=2,
-            steps=1,
-            loss=invalid_loss
-        )
-    assert "Custom loss function must accept exactly two arguments" in str(exc_info.value)
-
-def test_compile_kwargs_with_custom_loss():
-    """Test that custom loss takes priority over compile_kwargs"""
-    def custom_loss(y_true, y_pred):
-        return tf.reduce_mean(tf.square(y_true - y_pred))
-    
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    with pytest.warns(UserWarning):
-        model = create_and_compile_model(
-            series=data,
-            lags=2,
-            steps=1,
-            loss=custom_loss,
-            compile_kwargs={'loss': 'mse'}
-        )
-    assert model.loss == custom_loss
-
-def test_loss_and_optimizer_precedence():
-    """Test that main parameters take precedence over compile_kwargs"""
-    data = pd.DataFrame({'col1': [1, 2, 3], 'col2': [4, 5, 6]})
-    with pytest.warns(UserWarning, match="Using the optimizer parameter instead"):
-        with pytest.warns(UserWarning, match="Using the loss parameter instead"):
-            model = create_and_compile_model(
-                series=data,
-                lags=2,
-                steps=1,
-                optimizer="adam",
-                loss="mse",
-                compile_kwargs={'optimizer': 'sgd', 'loss': 'mae'}
-            )
-    assert model.loss == "mse"
